requirements.txt
tensorflow
opencv-python
numpy
streamlit
Pillow
h5py

